+++ src/modules/turmas/disciplina/disciplina.service.ts
@@ -190,80 +190,6 @@ export class DisciplinaService {
     }
   }
 
-  async removeCursoDisciplina({ codigo, nomeCurso }: DisciplinaDTO) {
-    const disciplinaExists = await this.prisma.disciplina.findUnique({
-      where: {
-        codigo,
-      },
-    });
-
-    if (!disciplinaExists) {
-      throw new AppError('Disciplina não existe');
-    }
-
-    const cursoExists = await this.prisma.curso.findUnique({
-      where: {
-        nome: nomeCurso,
-      },
-    });
-
-    if (!cursoExists) {
-      throw new AppError('Curso não existe');
-    }
-
-    const disciplina = await this.prisma.disciplina.update({
-      where: {
-        codigo: codigo,
-      },
-      data: {
-        Curso: {
-          disconnect: {
-            nome: nomeCurso,
-          },
-        },
-      },
-    });
-
-    return disciplina;
-  }
-
-  async addCursoDisciplina({ codigo, nomeCurso }: DisciplinaDTO) {
-    const disciplinaExists = await this.prisma.disciplina.findUnique({
-      where: {
-        codigo,
-      },
-    });
-
-    if (!disciplinaExists) {
-      throw new AppError('Disciplina não existe');
-    }
-
-    const cursoExists = await this.prisma.curso.findUnique({
-      where: {
-        nome: nomeCurso,
-      },
-    });
-
-    if (!cursoExists) {
-      throw new AppError('Curso não existe');
-    }
-
-    const disciplina = await this.prisma.disciplina.update({
-      where: {
-        codigo: codigo,
-      },
-      data: {
-        Curso: {
-          connect: {
-            nome: nomeCurso,
-          },
-        },
-      },
-    });
-
-    return disciplina;
-  }
-
   async addDisciplinaAoAluno({ username, disciplinaId }: DisciplinaAlunoDTO) {
     const usuario = await this.prisma.usuario.findUnique({
       where: {
